pyproject.toml
[project]
name = "multi-agent-system"
version = "0.1.0"
description = "Multi-agent system using Strands Agents with Agent as Tools pattern"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=0.1.6",
    "strands-agents[openai]>=0.1.6",
    "strands-agents-tools>=0.1.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.45.1",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
